refactored upcoming_courses; dropdowns are hidden when user clicks anywhere on the webpage'
