+++ src/components/upcoming_courses/dummyData.js
@@ -1,0 +1,122 @@
+import months from "./months";
+const courseData = [
+  {
+    instructorImage:
+      "https://robohash.org/rationedolorevitae.jpg?size=80x80&set=set1",
+    instructorName: "Alys Coryndon",
+    courseName:
+      "Cras in purus eu magna vulputate luctus. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Vivamus vestibulum sagittis sapien. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Etiam vel augue. Vestibulum rutrum rutrum neque. Aenean auctor gravida sem.",
+    location:
+      "Proin leo odio, porttitor id, consequat in, consequat ut, nulla. Sed accumsan felis. Ut at dolor quis odio consequat varius.",
+    date: "10/8/2021",
+    time: "9:14 PM",
+    partnerName: "Photobean",
+    month: months[parseInt("10") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/pariaturnecessitatibuset.jpg?size=80x80&set=set1",
+    instructorName: "Raj Heda",
+    courseName:
+      "Nam ultrices, libero non mattis pulvinar, nulla pede ullamcorper augue, a suscipit nulla elit ac nulla. Sed vel enim sit amet nunc viverra dapibus. Nulla suscipit ligula in lacus. Curabitur at ipsum ac tellus semper interdum.",
+    location:
+      "Aenean lectus. Pellentesque eget nunc. Donec quis orci eget orci vehicula condimentum.",
+    date: "9/21/2021",
+    time: "2:20 PM",
+    partnerName: "Kwideo",
+    month: months[parseInt("9") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/etametmolestiae.jpg?size=80x80&set=set1",
+    instructorName: "Raj Heda",
+    courseName:
+      "Pellentesque eget nunc. Donec quis orci eget orci vehicula condimentum. Curabitur in libero ut massa volutpat convallis. Morbi odio odio, elementum eu, interdum eu, tincidunt in, leo.",
+    location:
+      "Aenean lectus. Pellentesque eget nunc. Donec quis orci eget orci vehicula condimentum.",
+    date: "6/7/2021",
+    time: "2:59 PM",
+    partnerName: "Yacero",
+    month: months[parseInt("6") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/sedomnisfugit.jpg?size=80x80&set=set1",
+    instructorName: "Raj Heda",
+    courseName:
+      "Donec vitae nisi. Nam ultrices, libero non mattis pulvinar, nulla pede ullamcorper augue, a suscipit nulla elit ac nulla. Sed vel enim sit amet nunc viverra dapibus. Nulla suscipit ligula in lacus. Curabitur at ipsum ac tellus semper interdum. Mauris ullamcorper purus sit amet nulla. Quisque arcu libero, rutrum ac, lobortis vel, dapibus at, diam.",
+    location:
+      "Cras non velit nec nisi vulputate nonummy. Maecenas tincidunt lacus at velit. Vivamus vel nulla eget eros elementum pellentesque.\n\nQuisque porta volutpat erat. Quisque erat eros, viverra eget, congue eget, semper rutrum, nulla. Nunc purus.\n\nPhasellus in felis. Donec semper sapien a libero. Nam dui.",
+    date: "4/22/2021",
+    time: "5:45 AM",
+    partnerName: "Browsedrive",
+    month: months[parseInt("4") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/laborevoluptatibusaut.jpg?size=80x80&set=set1",
+    instructorName: "Alys Coryndon",
+    courseName:
+      "In hac habitasse platea dictumst. Aliquam augue quam, sollicitudin vitae, consectetuer eget, rutrum at, lorem. Integer tincidunt ante vel ipsum. Praesent blandit lacinia erat. Vestibulum sed magna at nunc commodo placerat. Praesent blandit.",
+    location:
+      "Quisque id justo sit amet sapien dignissim vestibulum. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Nulla dapibus dolor vel est. Donec odio justo, sollicitudin ut, suscipit a, feugiat et, eros.\n\nVestibulum ac est lacinia nisi venenatis tristique. Fusce congue, diam id ornare imperdiet, sapien urna pretium nisl, ut volutpat sapien arcu sed augue. Aliquam erat volutpat.",
+    date: "10/9/2021",
+    time: "9:55 AM",
+    partnerName: "Voomm",
+    month: months[parseInt("10") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/omnisametnobis.jpg?size=80x80&set=set1",
+    instructorName: "Alys Coryndon",
+    courseName:
+      "Vivamus in felis eu sapien cursus vestibulum. Proin eu mi. Nulla ac enim. In tempor, turpis nec euismod scelerisque, quam turpis adipiscing lorem, vitae mattis nibh ligula nec sem.",
+    location:
+      "Phasellus sit amet erat. Nulla tempus. Vivamus in felis eu sapien cursus vestibulum.\n\nProin eu mi. Nulla ac enim. In tempor, turpis nec euismod scelerisque, quam turpis adipiscing lorem, vitae mattis nibh ligula nec sem.\n\nDuis aliquam convallis nunc. Proin at turpis a pede posuere nonummy. Integer non velit.",
+    date: "5/8/2021",
+    time: "4:23 PM",
+    partnerName: "Jetwire",
+    month: months[parseInt("5") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/rerumdoloret.jpg?size=80x80&set=set1",
+    instructorName: "Alys Coryndon",
+    courseName:
+      "Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Donec pharetra, magna vestibulum aliquet ultrices, erat tortor sollicitudin mi, sit amet lobortis sapien sapien non mi. Integer ac neque.",
+    location:
+      "Donec diam neque, vestibulum eget, vulputate ut, ultrices vel, augue. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Donec pharetra, magna vestibulum aliquet ultrices, erat tortor sollicitudin mi, sit amet lobortis sapien sapien non mi. Integer ac neque.",
+    date: "11/20/2021",
+    time: "12:51 PM",
+    partnerName: "Leenti",
+    month: months[parseInt("11") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/placeatcumdolorum.jpg?size=80x80&set=set1",
+    instructorName: "Claribel Drewell",
+    courseName:
+      "Vivamus metus arcu, adipiscing molestie, hendrerit at, vulputate vitae, nisl. Aenean lectus. Pellentesque eget nunc. Donec quis orci eget orci vehicula condimentum. Curabitur in libero ut massa volutpat convallis. Morbi odio odio, elementum eu, interdum eu, tincidunt in, leo. Maecenas pulvinar lobortis est.",
+    location:
+      "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Proin risus. Praesent lectus.",
+    date: "10/1/2021",
+    time: "3:23 PM",
+    partnerName: "Zoomdog",
+    month: months[parseInt("10") - 1],
+  },
+  {
+    instructorImage:
+      "https://robohash.org/nulladucimusmagni.jpg?size=80x80&set=set1",
+    instructorName: "Claribel Drewell",
+    courseName:
+      "Vivamus metus arcu, adipiscing molestie, hendrerit at, vulputate vitae, nisl. Aenean lectus. Pellentesque eget nunc. Donec quis orci eget orci vehicula condimentum.",
+    location:
+      "Quisque id justo sit amet sapien dignissim vestibulum. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Nulla dapibus dolor vel est. Donec odio justo, sollicitudin ut, suscipit a, feugiat et, eros.\n\nVestibulum ac est lacinia nisi venenatis tristique. Fusce congue, diam id ornare imperdiet, sapien urna pretium nisl, ut volutpat sapien arcu sed augue. Aliquam erat volutpat.\n\nIn congue. Etiam justo. Etiam pretium iaculis justo.",
+    date: "11/19/2021",
+    time: "6:47 AM",
+    partnerName: "Gabcube",
+    month: months[parseInt("11") - 1],
+  },
+];
+
+export default courseData;
