+++ src/components/upcoming_courses/months.js
@@ -1,0 +1,16 @@
+const months = [
+  "january",
+  "february",
+  "march",
+  "april",
+  "may",
+  "june",
+  "july",
+  "august",
+  "september",
+  "october",
+  "november",
+  "december",
+];
+
+export default months;
